+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (DEV) Added array and object type to command line parameters.
 - `abeamer render` can pass `--render-var` parameters to client lib as `args.renderVars`.
 - `story.args` is exposed as readonly property to allow to access `renderVars`.
+- `gallery/animate-badges` reads parameters via `--render-vars`.
 
 ## [0.3.2] -
 ### Changed